refactor(auth): move user token routes
